pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic-settings = "^2.10.1"
python-multipart = "^0.0.20"
pypdf4 = "^1.27.0"
pymupdf = "^1.26.4"
translitua = "^1.3.2"
pyhumps = "^3.8.0"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

